# Various minor fixes