--- a/benchmarks/src/benchmark.cpp
+++ b/benchmarks/src/benchmark.cpp
@@ -1,4 +1,3 @@
-#include "unicode/ucnv.h"
 #include "benchmark.h"
 #include "simdutf.h"