--- a/src/haswell/implementation.cpp
+++ b/src/haswell/implementation.cpp
@@ -7,6 +7,8 @@
 #include "scalar/utf8.h"
 #include "scalar/utf16.h"
 #include "scalar/latin1.h"
+#include "scalar/utf8_to_latin1/valid_utf8_to_latin1.h"
+#include "scalar/utf8_to_latin1/utf8_to_latin1.h"
 
 #include "simdutf/haswell/begin.h"
 namespace simdutf {
@@ -80,9 +82,8 @@ simdutf_warn_unused size_t implementation::convert_latin1_to_utf16be(const char*
 }
 
 simdutf_warn_unused size_t implementation::utf16_length_from_latin1(const char * input, size_t length) const noexcept {
-  //return length;
-    return scalar::latin1::utf16_length_from_latin1(input,length);
-  }
+  return scalar::latin1::utf16_length_from_latin1(input,length);
+}
 
 simdutf_warn_unused size_t implementation::convert_valid_utf16be_to_latin1(const char16_t* buf, size_t len, char* latin1_output) const noexcept {
   return scalar::utf16_to_latin1::convert_valid<endianness::BIG>(buf, len, latin1_output);
@@ -104,14 +105,13 @@ simdutf_warn_unused result implementation::convert_utf16le_to_latin1_with_errors
   return scalar::utf16_to_latin1::convert_with_errors<endianness::LITTLE>(buf, len, latin1_output);
 }
 
-
 simdutf_warn_unused size_t implementation::convert_utf16le_to_latin1(const char16_t* buf, size_t len, char* latin1_output) const noexcept {
   return scalar::utf16_to_latin1::convert<endianness::LITTLE>(buf, len, latin1_output);
 }
 
 simdutf_warn_unused size_t implementation::latin1_length_from_utf16(const char16_t * input, size_t length) const noexcept {
   return scalar::utf16::latin1_length_from_utf16(input,length);
-  }
+}
 
 simdutf_warn_unused size_t implementation::convert_latin1_to_utf8(const char * buf, size_t len, char* utf8_output) const noexcept {
   return scalar::latin1_to_utf8::convert(buf,len,utf8_output);
@@ -120,8 +120,6 @@ simdutf_warn_unused size_t implementation::convert_latin1_to_utf8(const char * b
 simdutf_warn_unused size_t implementation::latin1_length_from_utf8(const char* buf, size_t len) const noexcept {
   return scalar::utf8::latin1_length_from_utf8(buf,len);
 }
-
-
 
 simdutf_warn_unused size_t implementation::convert_utf8_to_latin1(const char* buf, size_t len, char* latin1_output) const noexcept {
   return scalar::utf8_to_latin1::convert(buf, len, latin1_output);