I made a few minor changes. 

- [x]  Somehow the code would not build without additional headers. I added them.
- [x] There was an unguarded ICU header.
- [x] I fixed a few formatting issues. Though the simdutf project does not have compulsory formatting, we try to keep the code nice.

Here are my performance numbers:

```
$ ./build/benchmarks/benchmark -P convert_utf8_to_latin1  -F unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 67.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_latin1+haswell, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   3.004 ins/byte,    0.627 cycle/byte,    3.190 GB/s (1.6 %),     2.001 GHz,    4.789 ins/cycle 
   3.004 ins/char,    0.627 cycle/char,    3.190 Gc/s (1.6 %)     1.00 byte/char 
convert_utf8_to_latin1+icelake, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   3.004 ins/byte,    0.627 cycle/byte,    3.189 GB/s (1.5 %),     2.001 GHz,    4.787 ins/cycle 
   3.004 ins/char,    0.627 cycle/char,    3.189 Gc/s (1.5 %)     1.00 byte/char 
convert_utf8_to_latin1+iconv, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
  30.025 ins/byte,    5.254 cycle/byte,    0.608 GB/s (21.2 %),     3.194 GHz,    5.715 ins/cycle 
  30.025 ins/char,    5.254 cycle/char,    0.608 Gc/s (21.2 %)     1.00 byte/char 
WARNING: Measurements are noisy, try increasing iteration count (-I).
convert_utf8_to_latin1+icu, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
  17.171 ins/byte,    3.043 cycle/byte,    0.939 GB/s (0.5 %),     2.858 GHz,    5.643 ins/cycle 
  17.171 ins/char,    3.043 cycle/char,    0.939 Gc/s (0.5 %)     1.00 byte/char 
convert_utf8_to_latin1+westmere, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   3.004 ins/byte,    0.629 cycle/byte,    5.091 GB/s (0.9 %),     3.202 GHz,    4.776 ins/cycle 
   3.004 ins/char,    0.629 cycle/char,    5.091 Gc/s (0.9 %)     1.00 byte/char 
convert_utf8_to_latin1_with_errors+haswell, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   3.004 ins/byte,    0.627 cycle/byte,    5.106 GB/s (0.8 %),     3.202 GHz,    4.790 ins/cycle 
   3.004 ins/char,    0.627 cycle/char,    5.106 Gc/s (0.8 %)     1.00 byte/char 
convert_utf8_to_latin1_with_errors+icelake, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   3.004 ins/byte,    0.627 cycle/byte,    5.106 GB/s (0.8 %),     3.202 GHz,    4.790 ins/cycle 
   3.004 ins/char,    0.627 cycle/char,    5.106 Gc/s (0.8 %)     1.00 byte/char 
convert_utf8_to_latin1_with_errors+westmere, input size: 86940, iterations: 3000, dataset: unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
   3.004 ins/byte,    0.628 cycle/byte,    5.099 GB/s (0.6 %),     3.202 GHz,    4.784 ins/cycle 
   3.004 ins/char,    0.628 cycle/char,    5.099 Gc/s (0.6 %)     1.00 byte/char 
```

I have not investigated further but if I try to process non-ASCII files, I get errors as part of the benchmarking...

```
git clone git@github.com:lemire/unicode_lipsum.git
./build/benchmarks/benchmark -P latin -F unicode_lipsum/wikipedia_mars/esperanto.latin1.txt 
```

This does not work.